Refactor and style up FormSelect
To be able to include a custom select with stylable menu
in a quick manner, we implemented React Select.

As a compromise between using React Select and A11Y, we styled
FormSelect as a hybrid component, with a <select> for non-visual access
and keyboard handling, and React Select for mouse handling.